JRM, progres

+++ final/library/src/Queue.cpp
@@ -3,7 +3,7 @@
  * Name: Queue.cpp
  * Description: Random access circular buffer data structure
  * 
- * Version: 1.0
+ * Version: 1.1
 \******************************************************************************/
 
 #include "Queue.h"
@@ -11,7 +11,7 @@
 Queue::Queue()
 : head_(0) {}
 
-void Queue::push(uint8_t data) {
+void Queue::push(uint16_t data) {
 	
 	// Overwrite the oldest data
 	data_[head_] = data;
@@ -26,7 +26,7 @@ void Queue::push(uint8_t data) {
 
 }
 
-uint8_t Queue::get(uint8_t i) const {
+uint16_t Queue::get(uint8_t i) const {
 	// Access the data n+1 positions behind head
 	// Ensures the index is within bounds
 	return data_[(head_-(i+1)+QUEUE_CAPACITY) % QUEUE_CAPACITY];
